converting files to typescript
