+++ backend/notes/views.py
@@ -30,11 +30,11 @@ def createNote(request):
         created = dateStamp,
         updated = dateStamp
     ))
-    
+
     # -- check if data is without bad actors
     if noteSerializer.is_valid():
         noteSerializer.save()
-        
+
         return Response(data=noteSerializer.data,status=status.HTTP_201_CREATED)
     else:
         return Response(data=noteSerializer.errors,status=status.HTTP_400_BAD_REQUEST)
@@ -44,7 +44,7 @@ def createNote(request):
 @api_view(['GET'])
 def myNotes(request):
     user = getUserID(request)
-    
+
     try:
         return Response(data=NoteListSerializer(Note.objects.filter(user=user).values('id','user','created','updated','title').order_by('updated'), many=True).data, status=status.HTTP_200_OK)
     except Note.DoesNotExist:
@@ -66,7 +66,7 @@ def noteOps(request, noteid):
 # Read
 def readNote(request, noteid):
     user = getUserID(request)
-    
+
     try:
         return Response(data=NoteSerializer(Note.objects.get(id=noteid, user=user)).data, status=status.HTTP_200_OK)
     except Note.DoesNotExist:
@@ -75,7 +75,7 @@ def readNote(request, noteid):
 # Delete
 def deleteNote(request, noteid):
     user = getUserID(request)
-    
+
     try:
         Note.objects.get(id=noteid, user=user).delete()
         return Response(data=successResponse(), status=status.HTTP_200_OK)
@@ -84,7 +84,7 @@ def deleteNote(request, noteid):
 
 # Update
 def updateNote(request, noteid):
-    user = getUserID(request)   
+    user = getUserID(request)
     try:
         note = Note.objects.get(id=noteid, user=user)
         note.title = request.data['title']
@@ -95,17 +95,17 @@ def updateNote(request, noteid):
         return Response(data=NoteSerializer(note).data,status=status.HTTP_200_OK)
     except Note.DoesNotExist:
         return Response(data=errorResponse("This note does not exist.","N0403"),status=status.HTTP_400_BAD_REQUEST)
-    
+
 # Share a note via unique URL
 # -----------------------------------------------
 # Create a URL
 @api_view(['POST'])
 def createNoteShareExternal(request, noteid):
     user = getUserID(request)
-    
+
     # Check if the user requesting is the owner
     # try:
-    if user == Note.objects.get(id=noteid).user: 
+    if user == Note.objects.get(id=noteid).user:
         rid = uuid.uuid4()
         expire = request.data['expire'] if type(request.data['expire']) is int else 0
         title = request.data['title']
@@ -123,20 +123,20 @@ def createNoteShareExternal(request, noteid):
             anonymous = anon
         )
         row.save()
-        
+
         return Response(data=dict(title=title, urlkey=rid, expire=expire, anonymous=anon, id=dbid, created=created), status=status.HTTP_201_CREATED)
     # except:
     #     return Response(data=errorResponse("The request could not be processed.","N1012"), status=status.HTTP_400_BAD_REQUEST)
-        
+
     return Response(data=errorResponse("The request is invalid.","N1001"), status=status.HTTP_400_BAD_REQUEST)
-    
+
 # Read the note
 @api_view(['GET'])
 def readNoteShareExternal(request, nui, permkey):
     try:
         noteShareExternal = ShareExternal.objects.get(key=hashThis(permkey),noteid__startswith=nui)
         note = Note.objects.get(id=noteShareExternal.noteid)
-        
+
         # We keep person who externally shared the note anonymous
         response = dict(
             noteTitle = note.title,
@@ -151,16 +151,16 @@ def readNoteShareExternal(request, nui, permkey):
             user = User.objects.get(id=noteShareExternal.creator)
             response['noteSharedBy'] = user.full_name
             response['noteSharedByUID'] = user.id
-        
+
         return Response(data=response, status=status.HTTP_200_OK)
     except (ShareExternal.DoesNotExist, Note.DoesNotExist):
         return Response(data=errorResponse("This note does not exist.","N1404"), status=status.HTTP_404_NOT_FOUND)
-    
+
 # Read all share links for the note
 @api_view(['GET'])
 def readAllLinksNoteShareExt(request, noteid):
     user = getUserID(request)
-    links = ShareExternalSerializer(ShareExternal.objects.filter(noteid=noteid,creator=user).values('id','anonymous','created','title').order_by('created'), many=True).data
+    links = ShareExternalSerializer(ShareExternal.objects.filter(noteid=noteid,creator=user).values('id','anonymous','created','title').order_by('-created'), many=True).data
     return Response(data=links, status=status.HTTP_200_OK)
     # except ShareExternal.DoesNotExist:
     #     return Response(data=errorResponse("No share links for this note yet.", "N1410"), status=status.HTTP_404_NOT_FOUND)
