+++ frontend/package.json
@@ -1,63 +1,64 @@
 {
-  "name": "letsnote",
-  "version": "0.1.1",
-  "private": true,
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.9.1",
-    "@types/jest": "^29.5.1",
-    "@types/node": "^18.16.3",
-    "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.1",
-    "axios": "^1.1.3",
-    "deep-equal": "^2.0.5",
-    "editorjs-html": "^3.4.2",
-    "immer": "^9.0.18",
-    "is-hotkey": "^0.2.0",
-    "is-url": "^1.2.4",
-    "lodash": "^4.17.21",
-    "react": "^18.2.0",
-    "react-cookie": "^4.1.1",
-    "react-dom": "^18.2.0",
-    "react-editor-js": "^2.1.0",
-    "react-helmet": "^6.1.0",
-    "react-redux": "^8.0.4",
-    "react-router-dom": "^6.4.2",
-    "react-scripts": "5.0.1",
-    "redux": "^4.2.0",
-    "slate": "^0.84.0",
-    "slate-history": "^0.85.0",
-    "slate-react": "^0.83.2",
-    "typescript": "^4.9.5"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "css": "npx tailwindcss -i ./src/assets/css/tailwind.css -o ./src/assets/css/main.css",
-    "watch-css": "npx tailwindcss -i ./src/assets/css/tailwind.css -o ./src/assets/css/main.css --watch"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.12",
-    "postcss": "^8.4.18",
-    "tailwindcss": "^3.2.0"
-  }
+	"name": "letsnote",
+	"version": "0.2.1",
+	"private": true,
+	"dependencies": {
+		"@reduxjs/toolkit": "^1.9.1",
+		"@types/jest": "^29.5.1",
+		"@types/node": "^18.16.3",
+		"@types/react": "^18.2.0",
+		"@types/react-dom": "^18.2.1",
+		"axios": "^1.1.3",
+		"deep-equal": "^2.0.5",
+		"editorjs-html": "^3.4.2",
+		"immer": "^9.0.18",
+		"is-hotkey": "^0.2.0",
+		"is-url": "^1.2.4",
+		"lodash": "^4.17.21",
+		"react": "^18.2.0",
+		"react-cookie": "^4.1.1",
+		"react-dom": "^18.2.0",
+		"react-editor-js": "^2.1.0",
+		"react-helmet": "^6.1.0",
+		"react-redux": "^8.0.4",
+		"react-router-dom": "^6.4.2",
+		"react-scripts": "5.0.1",
+		"redux": "^4.2.0",
+		"slate": "^0.84.0",
+		"slate-history": "^0.85.0",
+		"slate-react": "^0.83.2",
+		"typescript": "^4.9.5"
+	},
+	"scripts": {
+		"start": "react-scripts start",
+		"build": "react-scripts build",
+		"test": "react-scripts test",
+		"eject": "react-scripts eject",
+		"css": "npx tailwindcss -i ./src/assets/css/tailwind.css -o ./src/assets/css/main.css",
+		"watch-css": "npx tailwindcss -i ./src/assets/css/tailwind.css -o ./src/assets/css/main.css --watch"
+	},
+	"eslintConfig": {
+		"extends": [
+			"react-app",
+			"react-app/jest"
+		]
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"devDependencies": {
+		"@types/react-helmet": "^6.1.6",
+		"autoprefixer": "^10.4.12",
+		"postcss": "^8.4.18",
+		"tailwindcss": "^3.2.0"
+	}
 }
