+++ frontend/src/assets/css/editor.css
@@ -1,0 +1,203 @@
+/*
+  * Editor
+*/
+
+.ab-toolbar {
+	@apply w-full max-w-6xl mx-auto;
+}
+
+.editor {
+	position: relative;
+	font-family: "Roboto";
+	@apply mt-8;
+}
+
+.editor > div[role="textbox"] {
+	@apply max-w-4xl mb-10 lg:mb-16 px-6 py-4 sm:px-8 sm:py-6 sm:pb-36 lg:px-16 lg:py-12 lg:pb-48 xl:px-20 xl:py-16 xl:pb-64 sm:rounded-md shadow-sm bg-white mx-auto min-h-[calc(100vh-7rem)];
+}
+
+.editor a[data-slate-node] {
+	@apply text-sky-700 border-b border-solid border-sky-700 hover:text-blue-700 font-medium;
+}
+
+.editor code {
+	font-size: 0.94rem;
+	font-family: "Source Code Pro", "Consolas", monospace;
+	@apply bg-gray-300/80 text-gray-700 tracking-tight rounded px-1 mx-px py-px font-bold;
+}
+
+/*
+ * Ordered and Unoredered lists
+*/
+.editor-ul {
+	@apply list-disc;
+}
+.editor-ol {
+	@apply list-decimal;
+}
+
+.editor-ul,
+.editor-ol {
+	@apply ml-4 pb-1;
+}
+
+.editor > div[role="textbox"] > .editor-ul,
+.editor > div[role="textbox"] > .editor-ol {
+	@apply ml-8 my-2.5;
+}
+
+.editor-ul > .editor-li,
+.editor-ol > .editor-li {
+	@apply my-2 pl-2;
+}
+
+.editor-ol .editor-ol {
+	list-style: lower-alpha;
+}
+.editor-ol .editor-ol .editor-ol {
+	list-style: lower-roman;
+}
+
+.editor-ul .editor-ul {
+	list-style: circle;
+}
+.editor-ul .editor-ul .editor-ul {
+	list-style: square;
+}
+
+/*
+ * Block Elements
+*/
+
+.editor-h1 {
+	@apply mt-5 mb-5 text-4xl font-serif font-black;
+}
+
+.editor-h2 {
+	@apply mt-4 mb-4 text-[1.8rem] leading-7 font-serif font-black;
+}
+
+.editor-h3 {
+	@apply mt-3 mb-2.5 text-[1.5rem] leading-7 font-serif;
+}
+
+.editor-h4 {
+	@apply mt-3 mb-2.5 text-[1.2rem] leading-6 font-serif;
+}
+
+.editor-h2 + .editor-h3,
+.editor-h2 + .editor-h4 {
+	@apply mt-4;
+}
+
+.editor-h3 + .editor-h4 {
+	@apply mt-4;
+}
+.editor-highlight {
+	@apply bg-yellow-100;
+}
+
+.editor-p {
+	@apply block my-2.5 text-gray-900;
+}
+
+.editor-p,
+li {
+	line-height: 1.6rem;
+	font-size: 1rem;
+}
+
+.editor-p > span {
+	display: inline;
+}
+
+.editor-p + .editor-p {
+	@apply mt-2;
+}
+
+.editor-codeblock {
+	@apply bg-gray-100 bg-opacity-90 py-0.5 text-gray-700 font-medium text-sm px-2.5 font-mono border-l-2 border-gray-300 border-solid;
+}
+
+.editor-quote {
+	@apply border-l-2 border-gray-300 text-gray-600 px-4 my-2 font-normal italic;
+}
+
+.editor-codeblock + .editor-quote,
+.editor-p + .editor-quote,
+.editor-h1 + .editor-quote,
+.editor-h2 + .editor-quote,
+.editor-h3 + .editor-quote,
+.editor-h4 + .editor-quote {
+	@apply mt-3.5;
+}
+
+.editor-codeblock + .editor-p {
+	margin-top: 1.125rem;
+}
+
+.editor-quote + .editor-codeblock,
+.editor-h1 + .editor-codeblock,
+.editor-h2 + .editor-codeblock,
+.editor-h3 + .editor-codeblock,
+.editor-h4 + .editor-codeblock,
+.editor-p + .editor-codeblock {
+	@apply pt-0.5;
+}
+
+.editor-codeblock + .editor-h2,
+.editor-codeblock + .editor-h1,
+.editor-codeblock + .editor-h3,
+.editor-codeblock + .editor-h4 {
+	@apply mt-8;
+}
+
+.editor-p + .editor-h2,
+.editor-p + .editor-h1 {
+	@apply mt-6;
+}
+
+.editor-p + .editor-h3 {
+	@apply mt-6;
+}
+
+.editor-p + .editor-h4 {
+	@apply mt-4;
+}
+
+.editor-h1 + .editor-p,
+.editor-h2 + .editor-p {
+	@apply mt-3;
+}
+
+.editor-h3 + .editor-p {
+	@apply mt-2.5;
+}
+
+.editor-h4 + .editor-p {
+	@apply mt-2;
+}
+.editor-quote + .editor-codeblock {
+	@apply mt-3;
+}
+
+.editor-h1 + .editor-codeblock,
+.editor-h2 + .editor-codeblock,
+.editor-h3 + .editor-codeblock,
+.editor-h4 + .editor-codeblock {
+	@apply mt-4;
+}
+
+.editor-h1 + .editor-ul,
+.editor-h2 + .editor-ul,
+.editor-h3 + .editor-ul,
+.editor-h1 + .editor-ol,
+.editor-h2 + .editor-ol,
+.editor-h3 + .editor-ol {
+	@apply mt-3;
+}
+
+.editor-h4 + .editor-ul,
+.editor-h4 + .editor-ol {
+	@apply mt-2.5;
+}
