+++ frontend/src/components/home/sidebar/Sidebar.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+export interface ISidebarProps {}
+
+export function Sidebar(props: ISidebarProps) {
+	return <div></div>;
+}
