--- frontend/src/config.js
@@ -1,2 +1,0 @@
-export const BACKEND_SERVER_DOMAIN = "http://localhost:8000";
-// export const BACKEND_SERVER_DOMAIN = "https://letsnote.io";
