+++ frontend/src/index.tsx
@@ -1,10 +1,8 @@
-import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import "./assets/css/main.css";
+// CSS
+import "./assets/css/global.css";
 import "./assets/css/icons.css";
 
-const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLInputElement
-);
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLInputElement);
 root.render(<App />);
