+++ frontend/src/lib/user/register.ts
@@ -1,0 +1,1 @@
+export {};
