--- frontend/src/utils/ExampleDocument.js
@@ -1,12 +1,0 @@
-const ExampleDocument = [
-  {
-    type: "paragraph",
-    children: [
-      {
-        text: "",
-      },
-    ],
-  },
-];
-
-export default ExampleDocument;
