+++ frontend/src/utils/ExampleDocument.ts
@@ -1,0 +1,19 @@
+const ExampleDocument: SlateNodeType[] = [
+	{
+		type: "paragraph",
+		children: [
+			{
+				text: "",
+			},
+		],
+	},
+];
+
+export interface SlateNodeType {
+	type: string;
+	children: {
+		text: string;
+	}[];
+}
+
+export default ExampleDocument;
