+++ frontend/tailwind.config.js
@@ -1,8 +1,19 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ["./src/**/*.{js,jsx,ts,tsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
+	content: ["./src/**/*.{js,jsx,ts,tsx}"],
+	theme: {
+		extend: {
+			fontFamily: {
+				serif: [
+					'"Merriweather"',
+					"Merriweather",
+					"Times New Roman",
+					"Georgia",
+					"serif",
+				],
+				sans: ['"DM Sans"', "DM Sans", "Roboto", "Arial", "sans-serif"],
+			},
+		},
+	},
+	plugins: [],
 };
